backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.23.3
solders==0.20.0
solana==0.32.0
python-multipart==0.0.6
python-json-logger==2.0.7
python-dotenv==1.0.0
